Moved tasks.js
