+++ src/__tests__/tasks.js
@@ -1,0 +1,24 @@
+const tasks = [
+  {
+    description: 'work1',
+    completed: true,
+    index: 1,
+  },
+  {
+    description: 'work2',
+    completed: false,
+    index: 2,
+  },
+  {
+    description: 'work3',
+    completed: false,
+    index: 3,
+  },
+  {
+    description: 'work4',
+    completed: true,
+    index: 4,
+  },
+];
+
+export default tasks;
